Make deb signing optional
I believe that with apt we only need to sign the meta files, not the
actual deb files for apt to be happy.

This fixes the issue where a deb that has been signed (e.g. 3rd party
download from a website) can't be uploaded (as you can't sign a signed
deb).